// Frontend/ShoppingCart/src/app/models/Cart/ICartItemCollectionModel.ts
/**
 * Defines the cart item summary details for each product.
 */
export interface ICartItemCollectionModel {
  /// <summary>
  /// Uniquely identifies each cart item with a unique order id
  /// </summary>
  /// <example>1</example>
  OrderID: number;
  /// <summary>
  /// Represents a unique product added to the cart
  /// </summary>
  /// <example>1</example>
  ProductID: number;
  /// <summary>
  /// Product name of the cart item
  /// </summary>
  /// <example>Branded Foods</example>
  Productname: string;
  /// <summary>
  /// Total quantity ordered by the user
  /// </summary>
  /// <example>2</example>
  Quantity: number;
  /// <summary>
  /// Defines the total price for this product for the ordered quantity.
  /// </summary>
  /// <example>300</example>
  TotalAmount: number;
  /// <summary>
  /// Defines the image name of this product which is autogenerated by the server
  /// </summary>
  /// <example>1.png</example>
  ProductImageName: string;
  /// <summary>
  /// Defines the image API URL from the server
  /// </summary>
  /// <example>https://localhost:44398/images/1.png</example>
  ProductImageURL: string;
  /// <summary>
  /// Uniquely identifies each order with the user.
  /// </summary>
  /// <example>1</example>
  UserID: number;
  /// <summary>
  /// Username of the user.
  /// </summary>
  /// <example>karthik</example>
  Username: string;
}